rename: digest -> digests
