+++ client/consensus/babe/src/verification.rs
@@ -19,7 +19,7 @@ use schnorrkel::vrf::{VRFOutput, VRFProof};
 use sp_runtime::{traits::Header, traits::DigestItemFor};
 use sp_core::{Pair, Public};
 use sp_consensus_babe::{AuthoritySignature, SlotNumber, AuthorityIndex, AuthorityPair, AuthorityId};
-use sp_consensus_babe::digest::{PreDigest, CompatibleDigestItem};
+use sp_consensus_babe::digests::{PreDigest, CompatibleDigestItem};
 use sc_consensus_slots::CheckedHeader;
 use log::{debug, trace};
 use super::{find_pre_digest, babe_err, Epoch, BlockT, Error};
