+++ primitives/consensus/babe/src/lib.rs
@@ -19,13 +19,13 @@
 #![forbid(unsafe_code, missing_docs, unused_variables, unused_imports)]
 #![cfg_attr(not(feature = "std"), no_std)]
 
-pub mod digest;
+pub mod digests;
 pub mod inherents;
 
 use codec::{Encode, Decode};
 use sp_std::vec::Vec;
 use sp_runtime::{ConsensusEngineId, RuntimeDebug};
-use crate::digest::NextEpochDescriptor;
+use crate::digests::NextEpochDescriptor;
 
 mod app {
 	use sp_application_crypto::{app_crypto, key_types::BABE, sr25519};
